Search multiple paths when loading deferred component .so files. (#23849)
This is a partial revert of 9e6f4fc281b3ed83012a5b544be57db2e39cf6da

On some devices we still need to the original search paths approach
because dlopen with just the base file name doesn't work. We're
combining both approaches now, adding the base filename as the first
entry in the searchPaths.